+++ src/Sulu/Bundle/WebsiteBundle/Entity/Analytics.php
@@ -79,7 +79,7 @@ class Analytics
      *
      * @param string $title
      *
-     * @return Analytics
+     * @return self
      */
     public function setTitle($title)
     {
@@ -103,7 +103,7 @@ class Analytics
      *
      * @param bool $allDomains
      *
-     * @return Analytics
+     * @return self
      */
     public function setAllDomains($allDomains)
     {
@@ -127,7 +127,7 @@ class Analytics
      *
      * @param string $content
      *
-     * @return Analytics
+     * @return self
      */
     public function setContent($content)
     {
@@ -151,7 +151,7 @@ class Analytics
      *
      * @param string $type
      *
-     * @return Analytics
+     * @return self
      */
     public function setType($type)
     {
@@ -175,7 +175,7 @@ class Analytics
      *
      * @param string $webspaceKey
      *
-     * @return Analytics
+     * @return self
      */
     public function setWebspaceKey($webspaceKey)
     {
@@ -199,7 +199,7 @@ class Analytics
      *
      * @param Domain $domain
      *
-     * @return Analytics
+     * @return self
      */
     public function addDomain(Domain $domain)
     {