+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
+    * ENHANCEMENT #XXXX [WebsiteBundle]           Analytics: Added possibility to determine the position of content (head open, head close, body open, body close)
     * HOTFIX     #3610 [ContactBundle]           ContactSelectionContentType: Fixed bug in serialization with missing category data
     * HOTFIX     #3597 [WebsiteBundle]           Only display analytics in the actual environment
     * HOTFIX     #3598 [SecurityBundle]          Fixed misleading security warning message